refactor: tobe 패키지 제거 및 mvc 패키지로 이동
