+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/HandlerMapping.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet;
+package com.interface21.webmvc.servlet.mvc;
 
 import java.io.FileNotFoundException;
