+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerKeyTest.java
@@ -3,6 +3,7 @@ package com.interface21.webmvc.servlet.mvc.tobe;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.interface21.web.bind.annotation.RequestMethod;
+import com.interface21.webmvc.servlet.mvc.HandlerKey;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
